fix: rename infra gemini
