+++ app/infra/gemini/get_answer.go
@@ -1,4 +1,4 @@
-package genemi
+package gemini
 
 import (
 	"context"
